requirements.txt
# Core Dependencies
streamlit==1.48.0
pandas==2.2.1
plotly==6.2.0
numpy==1.26.4
tenacity==8.5.0
python-dotenv==1.1.1
tiktoken==0.11.0

# Cross-platform user data dirs
platformdirs>=4.0.0


# Google API Dependencies
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
protobuf==4.25.8
requests==2.32.4

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
responses>=0.22.0
freezegun>=1.2.2
